pyproject.toml
[tool.poetry]
name = "boto-task-django"
version = "0.1.0"
description = ""
authors = ["Seliverstov <seli@verstov.pro>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.1"
djangorestframework = "^3.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
